Fix SetSessionTicketKeys broken by GetConfigForClient
GetConfigForClient needs to clone the provided *tls.Config when setting
the cipher suites. Previously this was done once when initially called,
but this breaks the use of SetSessionTicketKeys, and generally
complicates setting up the *tls.Config.

Instead move the calls to Clone into the returned function. This has the
downside of requiring a clone for every connection and the increased
garbage thereof.